video/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
httpx==0.28.1
boto3==1.34.115
google-genai==1.21.1
agno==1.6.3

# Optional: performance-optimized parsers (pure Python fallback already in place)
pydantic-core==2.16.3
orjson==3.9.12
ujson==5.8.0
